useProfilePicture, useFollowers, useUser hooks created
