+++ frontend/main/src/pages/main-page/HomePage.tsx
@@ -7,6 +7,7 @@ import { Box, CircularProgress, Fab, Modal, Paper, Tooltip } from '@mui/material
 import PostContentBox from '../../components/home-page/PostContentBox'
 import AddIcon from '@mui/icons-material/Add'
 import CreatePostBox from '../../components/home-page/CreatePostBox'
+import useUser from '../../hooks/useUser'
 
 const fabStyling = {
     position: 'fixed',
@@ -35,7 +36,7 @@ const dateUpdatedDescending = (a: UserContents, b: UserContents): number => {
     else if (b.content.dateUpdated)
         return -1
     
-    return (a.content.dateUpdated ?? new Date()) > (b.content.dateUpdated ?? new Date()) ? -1 : (a.content.dateUpdated === b.content.dateUpdated ? 0 : 1)
+    return (a.content.dateUpdated ?? new Date()) > (b.content.dateUpdated ?? new Date()) ? 1 : (a.content.dateUpdated === b.content.dateUpdated ? 0 : -1)
 }
 
 export interface UserContents {
@@ -44,30 +45,13 @@ export interface UserContents {
 }
 
 export const HomePage = () => {
-    const [ user, setUser ] = useState<UserModel | null>(null)
+    const [ user ] = useUser()
     const [ contents, setContents ] = useState<UserContents[]> ([])
     const [ contentLoadMessage, setContentLoadMessage ] = useState<string | null>(null)
     const [ createPostOpen, setCreatePostOpen ] = useState<boolean>(false)
 
     useEffect(() => {
-        const getUser = async(jwt: string | null | undefined) => {
-            if (jwt) {
-                try {
-                    const jwtResponse = await _apiClient.verifyToken(jwt)
-                    const response = await _apiClient.userGET(jwtResponse.email)
-                    setUser(response)
-                }
-                catch {
-                    setUser(null)
-                }
-            }
-        }
-        getUser(localStorage.getItem('token'))
-
-    }, [])
-
-    useEffect(() => {
-        const getUsersFollowing = async(user: UserModel | null) => {
+        const getUsersFollowing = async(user: UserModel | undefined) => {
             if (user?.following) {
                     let currentUserContents: UserContents[] = []
                     for (let userFollowing of (user.following)) {
